Rename http-request.server.js
Matches our agreed convention of using kebab case.
